% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-Pi.r
\name{cTarget}
\alias{cTarget}
\alias{print.cTarget}
\title{Definition for S3 class \code{cTarget}}
\usage{
cTarget(priority, predictor)

\method{print}{cTarget}(x, ...)
}
\arguments{
\item{priority}{a data frame}

\item{predictor}{a data frame}

\item{x}{an object of class \code{cTarget}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{cTarget}
}
\description{
\code{cTarget} has 2 components: priority and predictor.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
cTarget(priority, predictor)
}
}
\keyword{S3}
\keyword{classes}
