% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-Pi.r
\name{eTarget}
\alias{eTarget}
\alias{print.eTarget}
\title{Definition for S3 class \code{eTarget}}
\usage{
eTarget(evidence, metag)

\method{print}{eTarget}(x, ...)
}
\arguments{
\item{evidence}{a data frame}

\item{metag}{an 'igraph' object}

\item{x}{an object of class \code{eTarget}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{eTarget}
}
\description{
\code{eTarget} has 2 components: evidence and metag.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
eTarget(evidence, metag)
}
}
\keyword{S3}
\keyword{classes}
