% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-Pi.r
\name{pPerf}
\alias{pPerf}
\alias{print.pPerf}
\title{Definition for S3 class \code{pPerf}}
\usage{
pPerf(PRS, AUROC, Fmax, ROC_perf, PR_perf, Pred_obj)

\method{print}{pPerf}(x, ...)
}
\arguments{
\item{PRS}{a data frame}

\item{AUROC}{a scalar}

\item{Fmax}{a scalar}

\item{ROC_perf}{a ROCR 'performance' object for ROC curve}

\item{PR_perf}{a ROCR 'performance' object for PR curve}

\item{Pred_obj}{a ROCR 'prediction' object for other performance
measures}

\item{x}{an object of class \code{pPerf}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{pPerf}
}
\description{
\code{pPerf} mush have following components: PRS, AUROC, Fmax,
ROC_perf, PR_perf, Pred_obj.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
pPerf(PRS, AUROC, Fmax, ROC_perf, PR_perf, Pred_obj)
}
}
\keyword{S3}
\keyword{classes}
