% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethod-Pi.r
\name{sGS}
\alias{sGS}
\alias{print.sGS}
\title{Definition for S3 class \code{sGS}}
\usage{
sGS(GSN, GSP, g)

\method{print}{sGS}(x, ...)
}
\arguments{
\item{GSN}{a vector}

\item{GSP}{a vector}

\item{g}{an 'igraph' object}

\item{x}{an object of class \code{sGS}}

\item{...}{other parameters}
}
\value{
an object of S3 class \code{sGS}
}
\description{
\code{sGS} mush have following components: GSN, GSP, g.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
sGS(GSN, GSP, g)
}
}
\keyword{S3}
\keyword{classes}
