% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xContour.r
\name{xContour}
\alias{xContour}
\title{Function to visualise a numeric matrix as a contour plot}
\usage{
xContour(data, main = "", xlab = "", ylab = "", key = "",
nlevels = 50,
colormap = c("darkblue-lightblue-lightyellow-darkorange", "bwr", "jet",
"gbr", "wyr", "br", "yr", "rainbow", "wb"), highlight = c("none",
"min", "max"), highlight.col = "white", x.label.cex = 0.95,
x.label.srt = 30, signature = FALSE, ...)
}
\arguments{
\item{data}{a numeric matrix for the contour plot}

\item{main}{an overall title for the plot}

\item{xlab}{a title for the x axis. If specified, it will override
'names(dimnames(data))[1]'}

\item{ylab}{a title for the y axis. If specified, it will override
'names(dimnames(data))[2]'}

\item{key}{a title for the key plot (on the right)}

\item{nlevels}{the number of levels to partition the input matrix
values. The same level has the same color mapped to}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{highlight}{how to highlight the point. It can be 'none' for no
highlight (by default), 'min' for highlighting the point with the
minimum value of the matrix, and 'max' for highlighting the point with
the maximum value of the matrix}

\item{highlight.col}{the highlight colors}

\item{x.label.cex}{the x-axis label size}

\item{x.label.srt}{the x-axis label angle (in degree from horizontal)}

\item{signature}{a logical to indicate whether the signature is
assigned to the plot caption. By default, it sets FALSE}

\item{...}{additional graphic parameters. For most parameters, please
refer to
\url{http://stat.ethz.ch/R-manual/R-devel/library/graphics/html/filled.contour.html}}
}
\value{
invisible
}
\description{
\code{xContour} is supposed to visualise a numeric matrix as a contour
plot.
}
\note{
none
}
\examples{
x <- y <- seq(-4*pi, 4*pi, len=10)
r <- sqrt(outer(x^2, y^2, "+"))
data <- cos(r^2)*exp(-r/(2*pi))
xContour(data)
#xContour(data, signature=TRUE)
}
\seealso{
\code{\link{xContour}}
}
