% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGSEAbarplot.r
\name{xGSEAbarplot}
\alias{xGSEAbarplot}
\title{Function to visualise GSEA results using a barplot}
\usage{
xGSEAbarplot(eGSEA, top_num = 10, displayBy = c("nes", "adjp", "fdr",
"pvalue"), FDR.cutoff = 0.05, bar.label = TRUE, bar.label.size = 3,
bar.color = "lightyellow-orange", bar.width = 0.8,
wrap.width = NULL, font.family = "sans", signature = TRUE)
}
\arguments{
\item{eGSEA}{an object of class "eGSEA"}

\item{top_num}{the number of the top terms (sorted according to FDR or
adjusted p-values). If it is 'auto', only the significant terms (see
below FDR.cutoff) will be displayed}

\item{displayBy}{which statistics will be used for displaying. It can
be "nes" for normalised enrichment score (by default), "adjp" or "fdr"
for adjusted p value (or FDR), "pvalue" for p value}

\item{FDR.cutoff}{FDR cutoff used to declare the significant terms. By
default, it is set to 0.05. This option only works when setting top_num
(see above) is 'auto'}

\item{bar.label}{logical to indicate whether to label each bar with
FDR. By default, it sets to true for bar labelling}

\item{bar.label.size}{an integer specifying the bar labelling text
size. By default, it sets to 3}

\item{bar.color}{either NULL or fill color names ('lightyellow-orange'
by default)}

\item{bar.width}{bar width. By default, 80% of the resolution of the
data}

\item{wrap.width}{a positive integer specifying wrap width of name}

\item{font.family}{the font family for texts}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xGSEAbarplot} is supposed to visualise GSEA results using a
barplot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
bp <- xGSEAbarplot(eGSEA, top_num="auto", displayBy="nes")
#pdf(file="GSEA_barplot.pdf", height=6, width=12, compress=TRUE)
print(bp)
#dev.off()
}
}
\seealso{
\code{\link{xPierGSEA}}
}
