% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGSEAconciser.r
\name{xGSEAconciser}
\alias{xGSEAconciser}
\title{Function to make GSEA results conciser by removing redundant terms}
\usage{
xGSEAconciser(eGSEA, cutoff = c(0.9, 0.5), verbose = TRUE)
}
\arguments{
\item{eGSEA}{an object of class "eGSEA"}

\item{cutoff}{a cutoff vector used to remove redundant terms. By
default, it has the first element 0.9 and the second element 0.5. It
means, for a term (less significant; called 'B'), if there is a more
significant term (called 'A'), their overlapped members cover at least
90% of the B's members, and also at least 50% of the A's members, then
this term B will be defined as redundant and thus being removed}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}
}
\value{
an object of class "eGSEA", after redundant terms being removed.
}
\description{
\code{xGSEAconciser} is supposed to make GSEA results conciser by
removing redundant terms. A redundant term (called 'B') is claimed if
its overlapped part (A&B) with a more significant term (called 'A')
meets both criteria: 1) |A&B| > 0.9*|B|; and 2) |A&B| > 0.5*|A|.
}
\note{
none
}
\examples{
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
eGSEA_concise <- xGSEAconciser(eGSEA)
}
}
\seealso{
\code{\link{xPierGSEA}}
}
