% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGSEAdotplot.r
\name{xGSEAdotplot}
\alias{xGSEAdotplot}
\title{Function to visualise GSEA results using dot plot}
\usage{
xGSEAdotplot(eGSEA, top = 1, priority.color = c("lightblue",
"darkblue"),
xlab = NULL, title = NULL, subtitle = c("leading", "enrichment",
"both"), clab = "5-star\\nratings", x.scale = c("normal", "sqrt",
"log"),
peak = TRUE, leading = FALSE, leading.size = 2, compact = FALSE,
font.family = "sans", signature = TRUE)
}
\arguments{
\item{eGSEA}{an object of class "eGSEA"}

\item{top}{the number of the top enrichments to be visualised.
Alternatively, the gene set names can be queried}

\item{priority.color}{a character vector for coloring priority scores}

\item{xlab}{the label for x-axis. If NULL, it is 'Target ranks'}

\item{title}{the title. If NULL, it is term name followed by the number
of its annotations}

\item{subtitle}{the subtitle. It can be used to show 'leading' info,
'enrichment' info or 'both'}

\item{clab}{the label for colorbar. By default, it is '5-star ratings'}

\item{x.scale}{how to transform the x scale. It can be "normal" for no
transformation, "sqrt" for square root transformation, and "log" for
log-based transformation}

\item{peak}{logical to indicate whether the peak location is shown}

\item{leading}{logical to indicate whether the leading targets are
texted}

\item{leading.size}{the size of leading targets' texts. It only works
when the parameter 'leading' is enabled}

\item{compact}{logical to indicate whether the compact/void theme is
used. If TRUE, axes and legend info will be hidden}

\item{font.family}{the font family for texts}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot" or a list of "ggplot" objects.
}
\description{
\code{xGSEAdotplot} is supposed to visualise GSEA results using dot
plot. It returns an object of class "ggplot" or a list of "ggplot"
objects.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
gp <- xGSEAdotplot(eGSEA, top=1)
#gp <- xGSEAdotplot(eGSEA, top=1, peak=FALSE, compact=TRUE, signature=FALSE)
gp

ls_gp <- xGSEAdotplot(eGSEA, top=1:4, signature=FALSE)
library(gridExtra)
grid.arrange(grobs=ls_gp, ncol=2)
}
}
\seealso{
\code{\link{xPierGSEA}}
}
