% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGSsimulator.r
\name{xGSsimulator}
\alias{xGSsimulator}
\title{Function to simulate gold standard negatives (GSN) given gold standard positives (GSP) and a gene network}
\usage{
xGSsimulator(GSP, population = NULL, network = c("STRING_medium",
"STRING_low", "STRING_high", "STRING_highest", "PCommonsUN_high",
"PCommonsUN_medium")[c(1, 6)], network.customised = NULL,
neighbor.order = 1, verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{GSP}{a vector containing Gold Standard Positives (GSP)}

\item{population}{a vector containing population space in which gold
standard negatives (GSN) will be considered. By default, it is NULL,
meaning genes in the network will be used instead}

\item{network}{the built-in network. Currently two sources of network
information are supported: the STRING database (version 10) and the
Pathways Commons database (version 7). STRING is a meta-integration of
undirect interactions from the functional aspect, while Pathways
Commons mainly contains both undirect and direct interactions from the
physical/pathway aspect. Both have scores to control the confidence of
interactions. Therefore, the user can choose the different quality of
the interactions. In STRING, "STRING_highest" indicates interactions
with highest confidence (confidence scores>=900), "STRING_high" for
interactions with high confidence (confidence scores>=700),
"STRING_medium" for interactions with medium confidence (confidence
scores>=400), and "STRING_low" for interactions with low confidence
(confidence scores>=150). For undirect/physical interactions from
Pathways Commons, "PCommonsUN_high" indicates undirect interactions
with high confidence (supported with the PubMed references plus at
least 2 different sources), "PCommonsUN_medium" for undirect
interactions with medium confidence (supported with the PubMed
references). By default, "STRING_medium" and "PCommonsUN_medium" are
used}

\item{network.customised}{an object of class "igraph". By default, it
is NULL. It is designed to allow the user analysing their customised
network data that are not listed in the above argument 'network'. This
customisation (if provided) has the high priority over built-in
network}

\item{neighbor.order}{an integer giving the order of the neighborhood.
By default, it is 1-order neighborhood}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a list with following components:
\itemize{
\item{\code{GSP}: a vector containing GSP after considering the
population space}
\item{\code{GSN}: a vector containing simulated GSN}
\item{\code{g}: an "igraph" object}
\item{\code{call}: the call that produced this result}
}
}
\description{
\code{xGSsimulator} is supposed to simulate gold standard negatives
(GSN) given gold standard positives (GSP) and an input gene network.
GSN targets are those after excluding GSP targets and their 1-order (by
default) neighbors in the gene network.
}
\note{
If multiple graphs are provided, they will be unionised for use.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
GS <- xGSsimulator(GSP, population,
network=c("STRING_medium","PCommonsUN_medium"),
RData.location=RData.location)
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xPredictROCR}},
\code{\link{xMLrandomforest}}
}
