% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLcompare.r
\name{xMLcompare}
\alias{xMLcompare}
\title{Function to visualise cross-validation performance against tuning parameters}
\usage{
xMLcompare(list_ML, metric = c("ROC", "Sens", "Spec"), xlab = NA,
xlimits = c(0.5, 1))
}
\arguments{
\item{list_ML}{a list of class "train" or "train.formula" objects
(resulting from caret::train)}

\item{metric}{the performance metric to plot. It can be one of 'ROC',
'Sens' (Sensitivity) and 'Spec' (Specificity)}

\item{xlab}{a title for the x axis}

\item{xlimits}{the limit for the x axis}
}
\value{
an object of class "ggplot"
}
\description{
\code{xMLcompare} is supposed to visualise cross-validation performance
against tuning parameters.
}
\note{
none
}
\examples{
\dontrun{
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
gp <- xMLcompare(ls_ML, xlimits=c(0.5,1))
}
}
\seealso{
\code{\link{xMLcompare}}
}
