% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLdensity.r
\name{xMLdensity}
\alias{xMLdensity}
\title{Function to visualise machine learning results using density plot}
\usage{
xMLdensity(pTarget, displayBy = c("All", "GS", "GSN", "GSP",
"Putative"),
x.scale = c("sqrt", "normal"), signature = TRUE)
}
\arguments{
\item{pTarget}{an object of class "pTarget"}

\item{displayBy}{which targets will be used for displaying. It can be
one of "GS" for gold standard targets, "GSN" for gold standard
negatives, "GSP" for gold standard positives, "Putative" for putative
targets (non-GS), "All" for all targets (by default)}

\item{x.scale}{how to transform the x scale. It can be "normal" for no
transformation, and "sqrt" for square root transformation (by default)}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xMLdensity} is supposed to visualise machine learning results
using density plot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
gp <- xMLdensity(pTarget, displayBy="All")
gp
}
}
\seealso{
\code{\link{xMLrandomforest}}
}

