% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLdotplot.r
\name{xMLdotplot}
\alias{xMLdotplot}
\title{Function to visualise machine learning results using dot plot}
\usage{
xMLdotplot(pTarget, displayBy = c("importance2fold", "roc2fold",
"fmax2fold",
"importance_accurancy", "importance_gini", "ROC", "Fmax"), signature =
TRUE)
}
\arguments{
\item{pTarget}{an object of class "pTarget"}

\item{displayBy}{which statistics will be used for displaying. It can
be either statistics across folds ("importance2fold" for predictor
importance, "roc2fold" for AUC in ROC, "fmax2fold" for F-max in
Precision-Recall curve) or overall statistics ("importance_accurancy"
for predictor importance measured by accuracy decrease,
"importance_gini" for predictor importance measured by Gini decrease,
"ROC" for AUC in ROC, "Fmax" for F-max in Precision-Recall curve)}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xMLdotplot} is supposed to visualise machine learning results
using dot plot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
gp <- xMLdotplot(pTarget, displayBy="importance_accurancy")
gp
}
}
\seealso{
\code{\link{xMLrandomforest}}
}
