% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLfeatureplot.r
\name{xMLfeatureplot}
\alias{xMLfeatureplot}
\title{Function to visualise/assess features used for machine learning}
\usage{
xMLfeatureplot(df_predictor, GSP, GSN, displayBy = c("boxplot", "ROC",
"Fmax"), font.family = "sans", ...)
}
\arguments{
\item{df_predictor}{a data frame containing genes (in rows) and
predictors (in columns), with their predictive scores inside it. This
data frame must has gene symbols as row names}

\item{GSP}{a vector containing Gold Standard Positive (GSP)}

\item{GSN}{a vector containing Gold Standard Negative (GSN)}

\item{displayBy}{which statistics will be used for displaying. It can
be either "boxplot" for features themselves, "ROC" for AUC in ROC,
"Fmax" for F-max in Precision-Recall curve)}

\item{font.family}{the font family for texts}

\item{...}{additional parameters. Please refer to 'lattice::bwplot' for
the complete list.}
}
\value{
an object of class "ggplot" for AUC and F-max, and an object of class
"trellis" for boxplot
}
\description{
\code{xMLfeatureplot} is supposed to visualise/assess features used for
machine learning. Visualisation can be made using either boxplot or dot
plot for AUC and F-max. It returns an object of class "ggplot" for AUC
and F-max, and an object of class "trellis" for boxplot.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xMLfeatureplot(df_predictor, GSP, GSN, displayBy="ROC")
}
}
