% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLparameters.r
\name{xMLparameters}
\alias{xMLparameters}
\title{Function to visualise cross-validation performance against tuning parameters}
\usage{
xMLparameters(data, nD = c("auto", "1D", "2D", "3D"), contour = TRUE,
main = "Repeated cross-validation", xlab = NA, ylab = NA, zlab = NA,
clab = "AUC (repeated CV)", nlevels = 50,
colormap = c("lightblue-lightyellow-darkorange-darkred", "bwr", "jet",
"gbr", "wyr", "br", "yr", "rainbow", "wb"), highlight = TRUE,
x.label.cex = 0.8, x.label.srt = 30, theta.3D = 40, phi.3D = 25)
}
\arguments{
\item{data}{an object of the class "train" or "train.formula"
(resulting from caret::train) used for 1D or 2D visualisation.
Alternatively, it can be a data frame used for 2D or 3D visualisation}

\item{nD}{an integer specifying the dimension of the visualisation. It
can be one of '1D', '2D' and '3D' and 'auto' (if input data is a
"train" object)}

\item{contour}{logical to indicate whether coutour plot should be also
included}

\item{main}{a title for the plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{zlab}{a title for the z axis}

\item{clab}{a title for the colorbar}

\item{nlevels}{the number of levels to partition the input matrix
values. The same level has the same color mapped to}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{highlight}{logical whether to highlight the point with the
maximum value}

\item{x.label.cex}{the x-axis label size}

\item{x.label.srt}{the x-axis label angle (in degree from horizontal)}

\item{theta.3D}{the azimuthal direction. By default, it is 40}

\item{phi.3D}{the colatitude direction. By default, it is 20}
}
\value{
invisible
}
\description{
\code{xMLparameters} is supposed to visualise cross-validation
performance against tuning parameters.
}
\note{
none
}
\examples{
\dontrun{
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
xMLparameters(df_fit, nD="2D")
xMLparameters(df_fit, nD="3D", theta.3D=40, phi.3D=60)
}
}
\seealso{
\code{\link{xMLparameters}}
}
