% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLrandomforest.r
\name{xMLrandomforest}
\alias{xMLrandomforest}
\title{Function to integrate predictor matrix in a supervised manner via machine learning algorithm random forest.}
\usage{
xMLrandomforest(df_predictor, GSP, GSN, nfold = 3, mtry = NULL,
ntree = 2000, verbose = TRUE, ...)
}
\arguments{
\item{df_predictor}{a data frame containing genes (in rows) and
predictors (in columns), with their predictive scores inside it. This
data frame must has gene symbols as row names}

\item{GSP}{a vector containing Gold Standard Positive (GSP)}

\item{GSN}{a vector containing Gold Standard Negative (GSN)}

\item{nfold}{an integer specifying the number of folds for cross
validataion}

\item{mtry}{an integer specifying the number of predictors randomly
sampled as candidates at each split. If NULL, it will be tuned by
`randomForest::tuneRF`, with starting value as sqrt(p) where p is the
number of predictors. The minimum value is 3}

\item{ntree}{an integer specifying the number of trees to grow. By
default, it sets to 2000}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}

\item{...}{additional graphic parameters. Please refer to
'randomForest::randomForest' for the complete list.}
}
\value{
an object of class "pTarget", a list with following components:
\itemize{
\item{\code{model}: a list of models, each results from per-fold train
set}
\item{\code{priority}: a data frame of nGene X 6 containing gene
priority information, where nGene is the number of genes in the input
data frame, and the 6 columns are "GS" (either 'GSP', or 'GSN', or
'Putative'), "name" (gene names), "rank" (ranks of the priority
scores), "pvalue" (the cross-fold aggregated p-value of being GSP,
per-fold p-value converted from empirical cumulative distribution of
the probability of being GSP), "fdr" (fdr adjusted from the aggregated
p-value), "priority" (-log10(fdr))}
\item{\code{predictor}: a data frame, which is the same as the input
data frame but inserting an additional column 'GS' in the first column}
\item{\code{prob2fold}: a data frame of nGene X 2+nfold containing the
probability of being GSP, where nGene is the number of genes in the
input data frame, nfold is the number of folds for cross validataion,
and the first two columns are "GS" (either 'GSP', or 'GSN', or
'Putative'), "name" (gene names), and the rest columns storing the
per-fold probability of being GSP}
\item{\code{importance2fold}: a data frame of nPredictor X 4+nfold
containing the predictor importance info, where nPredictor is the
number of predictors, nfold is the number of folds for cross
validataion, and the first 4 columns are "median" (the median of the
importance across folds), "mad" (the median of absolute deviation of
the importance across folds), "min" (the minimum of the importance
across folds), "max" (the maximum of the importance across folds), and
the rest columns storing the per-fold importance}
\item{\code{roc2fold}: a data frame of 1+nPredictor X 4+nfold
containing the integrated/predictor ROC info (AUC values), where
nPredictor is the number of predictors, nfold is the number of folds
for cross validataion, and the first 4 columns are "median" (the median
of the AUC values across folds), "mad" (the median of absolute
deviation of the AUC values across folds), "min" (the minimum of the
AUC values across folds), "max" (the maximum of the AUC values across
folds), and the rest columns storing the per-fold AUC values}
\item{\code{fmax2fold}: a data frame of 1+nPredictor X 4+nfold
containing the integrated/predictor PR info (F-max values), where
nPredictor is the number of predictors, nfold is the number of folds
for cross validataion, and the first 4 columns are "median" (the median
of the F-max values across folds), "mad" (the median of absolute
deviation of the F-max values across folds), "min" (the minimum of the
F-max values across folds), "max" (the maximum of the F-max values
across folds), and the rest columns storing the per-fold F-max values}
\item{\code{call}: the call that produced this result}
}
}
\description{
\code{xMLrandomforest} is supposed to integrate predictor matrix in a
supervised manner via machine learning algorithm random forest. It
requires three inputs: 1) Gold Standard Positive (GSP) targets; 2) Gold
Standard Negative (GSN) targets; 3) a predictor matrix containing genes
in rows and predictors in columns, with their predictive scores inside
it. It returns an object of class 'pTarget'.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
pTarget <- xMLrandomforest(df_prediction, GSP, GSN)
}
}

