% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLrename.r
\name{xMLrename}
\alias{xMLrename}
\title{Function to rename predictors used in machine learning}
\usage{
xMLrename(sTarget, old_names, new_names)
}
\arguments{
\item{sTarget}{an object of class "sTarget"}

\item{old_names}{a vector for the original names of predictors to be
renamed}

\item{new_names}{a vector for the new names}
}
\value{
an object of class "sTarget"
}
\description{
\code{xMLrename} is supposed to rename predictors used in machine
learning. It returns an object of class "sTarget".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
old_names <- colnames(sTarget$predictor)[-c(1,2)]
new_names <- c('nearbyGenes_nGene: nearby genes', 'eQTL_eGene: resting
state (CD14+)', 'eQTL_eGene: activating state (CD14+ by LPS2h)',
'eQTL_eGene: activating state (CD14+ by LPS24h)', 'eQTL_eGene:
activating state (CD14+ by IFN24h)', 'eQTL_eGene: B cells',
'eQTL_eGene: CD4+ T cells', 'eQTL_eGene: CD8+ T cells', 'eQTL_eGene:
neutrophils', 'eQTL_eGene: NK cells','eQTL_eGene: peripheral blood',
'HiC_cGene: monocytes', 'HiC_cGene: macrophages (M0)', 'HiC_cGene:
macrophages (M1)', 'HiC_cGene: macrophages (M2)', 'HiC_cGene:
neutrophils', 'HiC_cGene: CD4+ T cells (naive)', 'HiC_cGene: CD4+ T
cells (total)', 'HiC_cGene: CD8+ T cells (naive)', 'HiC_cGene: CD8+ T
cells (total)', 'HiC_cGene: B cells (naive)', 'HiC_cGene: B cells
(total)', 'Annotation_dGene: disease genes', 'Annotation_pGene:
phenotype genes', 'Annotation_fGene: function genes')
sTarget_rename <- xMLrename(sTarget, old_names, new_names)
}
}
\seealso{
\code{\link{xMLrandomforest}}
}
