% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLzoom.r
\name{xMLzoom}
\alias{xMLzoom}
\title{Function to visualise machine learning results using zoom plot}
\usage{
xMLzoom(xTarget, top = 20, top.label.type = c("box", "text"),
top.label.size = 3, top.label.query = NULL, point.shape = 3,
font.family = "sans", signature = TRUE)
}
\arguments{
\item{xTarget}{an object of class "sTarget" or "dTarget" (with the
component 'pPerf')}

\item{top}{the number of the top targets to be labelled/highlighted}

\item{top.label.type}{how to label the top targets. It can be "box"
drawing a box around the labels , and "text" for the text only}

\item{top.label.size}{the highlight label size}

\item{top.label.query}{which top genes in query will be labelled. By
default, it sets to NULL meaning all top genes will be displayed. If
labels in query can not be found, then none will be displayed}

\item{point.shape}{an integer specifying point shapes. By default, it
is 3 for cross. For details, please refere to
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/shape}}

\item{font.family}{the font family for texts}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE}
}
\value{
an object of class "ggplot"
}
\description{
\code{xMLzoom} is supposed to visualise machine learning results using
zoom plot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xMLzoom(sTarget)
gp
}
}
\seealso{
\code{\link{xMLrandomforest}}
}
