% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierABF.r
\name{xPierABF}
\alias{xPierABF}
\title{Function to prioritise genes based on seed eGenes identified through ABF integrating GWAS and eQTL summary data}
\usage{
xPierABF(
data,
eqtl = c("CD14", "LPS2", "LPS24", "IFN", "Bcell", "NK", "Neutrophil",
"CD4", "CD8",
"Blood", "Monocyte", "shared_CD14", "shared_LPS2", "shared_LPS24",
"shared_IFN"),
prior.eqtl = 1e-04,
prior.gwas = 1e-04,
prior.both = 1e-05,
cutoff.H4 = 0.8,
cutoff.pgwas = 1e-05,
network = c("STRING_highest", "STRING_high", "STRING_medium",
"STRING_low",
"PCommonsUN_high", "PCommonsUN_medium", "PCommonsDN_high",
"PCommonsDN_medium",
"PCommonsDN_Reactome", "PCommonsDN_KEGG", "PCommonsDN_HumanCyc",
"PCommonsDN_PID",
"PCommonsDN_PANTHER", "PCommonsDN_ReconX", "PCommonsDN_TRANSFAC",
"PCommonsDN_PhosphoSite", "PCommonsDN_CTD", "KEGG", "KEGG_metabolism",
"KEGG_genetic", "KEGG_environmental", "KEGG_cellular",
"KEGG_organismal",
"KEGG_disease"),
STRING.only = c(NA, "neighborhood_score", "fusion_score",
"cooccurence_score",
"coexpression_score", "experimental_score", "database_score",
"textmining_score")[1],
weighted = FALSE,
network.customised = NULL,
seeds.inclusive = TRUE,
normalise = c("laplacian", "row", "column", "none"),
restart = 0.7,
normalise.affinity.matrix = c("none", "quantile"),
parallel = TRUE,
multicores = NULL,
verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata",
guid = NULL
)
}
\arguments{
\item{data}{a data frame storing GWAS summary data with following
required columns 'snp', 'effect' (the effect allele assessed), 'other'
(other allele), 'b' (effect size for the allele assessed; log(odds
ratio) for a case-control study), 'se' (standard error), 'p' (p-value)}

\item{eqtl}{context-specific eQTL summary data. It can be one of
"Bcell","Blood","CD14","CD4","CD8","IFN","LPS24","LPS2","Monocyte","Neutrophil","NK","shared_CD14","shared_IFN","shared_LPS24","shared_LPS2"}

\item{prior.eqtl}{the prior probability an eQTL associated with the
eQTL trait. The default value is 1e-4}

\item{prior.gwas}{the prior probability an SNP associated with the GWAS
trait. The default value is 1e-4}

\item{prior.both}{the prior probability an eQTL/SNP associated with
both eQTL/GWAS traits. The default value is 1e-5}

\item{cutoff.H4}{the H4 cutoff used to define eGenes. This cutoff is
based on the posterior probabilities of H4 - one shared causal variant.
The default value is 0.8}

\item{cutoff.pgwas}{the GWAS p-value cutoff that must be met to
consider SNPs. The default value is 1e-5}

\item{network}{the built-in network. Currently two sources of network
information are supported: the STRING database (version 10) and the
Pathways Commons database (version 7). STRING is a meta-integration of
undirect interactions from the functional aspect, while Pathways
Commons mainly contains both undirect and direct interactions from the
physical/pathway aspect. Both have scores to control the confidence of
interactions. Therefore, the user can choose the different quality of
the interactions. In STRING, "STRING_highest" indicates interactions
with highest confidence (confidence scores>=900), "STRING_high" for
interactions with high confidence (confidence scores>=700),
"STRING_medium" for interactions with medium confidence (confidence
scores>=400), and "STRING_low" for interactions with low confidence
(confidence scores>=150). For undirect/physical interactions from
Pathways Commons, "PCommonsUN_high" indicates undirect interactions
with high confidence (supported with the PubMed references plus at
least 2 different sources), "PCommonsUN_medium" for undirect
interactions with medium confidence (supported with the PubMed
references). For direct (pathway-merged) interactions from Pathways
Commons, "PCommonsDN_high" indicates direct interactions with high
confidence (supported with the PubMed references plus at least 2
different sources), and "PCommonsUN_medium" for direct interactions
with medium confidence (supported with the PubMed references). In
addtion to pooled version of pathways from all data sources, the user
can also choose the pathway-merged network from individual sources,
that is, "PCommonsDN_Reactome" for those from Reactome,
"PCommonsDN_KEGG" for those from KEGG, "PCommonsDN_HumanCyc" for those
from HumanCyc, "PCommonsDN_PID" for those froom PID,
"PCommonsDN_PANTHER" for those from PANTHER, "PCommonsDN_ReconX" for
those from ReconX, "PCommonsDN_TRANSFAC" for those from TRANSFAC,
"PCommonsDN_PhosphoSite" for those from PhosphoSite, and
"PCommonsDN_CTD" for those from CTD. For direct (pathway-merged)
interactions sourced from KEGG, it can be 'KEGG' for all,
'KEGG_metabolism' for pathways grouped into 'Metabolism',
'KEGG_genetic' for 'Genetic Information Processing' pathways,
'KEGG_environmental' for 'Environmental Information Processing'
pathways, 'KEGG_cellular' for 'Cellular Processes' pathways,
'KEGG_organismal' for 'Organismal Systems' pathways, and 'KEGG_disease'
for 'Human Diseases' pathways}

\item{STRING.only}{the further restriction of STRING by interaction
type. If NA, no such restriction. Otherwide, it can be one or more of
"neighborhood_score","fusion_score","cooccurence_score","coexpression_score","experimental_score","database_score","textmining_score".
Useful options are c("experimental_score","database_score"): only
experimental data (extracted from BIND, DIP, GRID, HPRD, IntAct, MINT,
and PID) and curated data (extracted from Biocarta, BioCyc, GO, KEGG,
and Reactome) are used}

\item{weighted}{logical to indicate whether edge weights should be
considered. By default, it sets to false. If true, it only works for
the network from the STRING database}

\item{network.customised}{an object of class "igraph". By default, it
is NULL. It is designed to allow the user analysing their customised
network data that are not listed in the above argument 'network'. This
customisation (if provided) has the high priority over built-in
network. If the user provides the "igraph" object with the "weight"
edge attribute, RWR will assume to walk on the weighted network}

\item{seeds.inclusive}{logical to indicate whether non-network seed
genes are included for prioritisation. If TRUE (by default), these
genes will be added to the netowrk}

\item{normalise}{the way to normalise the adjacency matrix of the input
graph. It can be 'laplacian' for laplacian normalisation, 'row' for
row-wise normalisation, 'column' for column-wise normalisation, or
'none'}

\item{restart}{the restart probability used for Random Walk with
Restart (RWR). The restart probability takes the value from 0 to 1,
controlling the range from the starting nodes/seeds that the walker
will explore. The higher the value, the more likely the walker is to
visit the nodes centered on the starting nodes. At the extreme when the
restart probability is zero, the walker moves freely to the neighbors
at each step without restarting from seeds, i.e., following a random
walk (RW)}

\item{normalise.affinity.matrix}{the way to normalise the output
affinity matrix. It can be 'none' for no normalisation, 'quantile' for
quantile normalisation to ensure that columns (if multiple) of the
output affinity matrix have the same quantiles}

\item{parallel}{logical to indicate whether parallel computation with
multicores is used. By default, it sets to true, but not necessarily
does so. Partly because parallel backends available will be
system-specific (now only Linux or Mac OS). Also, it will depend on
whether these two packages "foreach" and "doMC" have been installed}

\item{multicores}{an integer to specify how many cores will be
registered as the multicore parallel backend to the 'foreach' package.
If NULL, it will use a half of cores available in a user's computer.
This option only works when parallel computation is enabled}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF
project. See \code{\link{xRDataLoader}} for details}
}
\value{
an object of class "pNode", a list with following components:
\itemize{
\item{\code{priority}: a matrix of nNode X 6 containing node priority
information, where nNode is the number of nodes in the input graph, and
the 5 columns are "name" (node names), "node" (1 for network genes, 0
for non-network seed genes), "seed" (1 for seeds, 0 for non-seeds),
"weight" (weight values), "priority" (the priority scores that are
rescaled to the range [0,1]), "rank" (ranks of the priority scores),
"description" (node description)}
\item{\code{g}: an input "igraph" object}
\item{\code{evidence}: a data frame storing evidence}
\item{\code{call}: the call that produced this result}
}
}
\description{
\code{xPierABF} is supposed to prioritise genes based on seed eGenes
identified through ABF integrating GWAS and eQTL summary data. To
prioritise genes, it first conducts colocalisation analysis through
Wakefield's Approximate Bayes Factor (ABF) integrating GWAS and eQTL
summary data to identify and score seed genes (that is, eGenes weighted
by posterior probability of the SNP being causal for both GWAS and eQTL
traits). It implements Random Walk with Restart (RWR) and calculates
the affinity score of all nodes in the graph to the seeds. The priority
score is the affinity score. Parallel computing is also supported for
Linux-like or Windows operating systems. It returns an object of class
"pNode".
}
\note{
The input graph will treat as an unweighted graph if there is no
'weight' edge attribute associated with
}
\examples{
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
data <- utils::read.delim(file="summary_gwas.RA.txt", header=TRUE,
row.names=NULL, stringsAsFactors=FALSE)
pNode_abf <- xPierABF(data, eqtl="Blood", network="STRING_high",
restart=0.7, RData.location=RData.location)
write.table(pNode_abf$priority, file="Genes_priority.ABF.txt",
sep="\t", row.names=FALSE)
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xMEabf}},
\code{\link{xPierGenes}}
}
