% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierABFheatmap.r
\name{xPierABFheatmap}
\alias{xPierABFheatmap}
\title{Function to visualise ABF evidence using heatmap}
\usage{
xPierABFheatmap(
data,
xTarget,
type = c("Gene", "Gene_SNP"),
colormap = "steelblue-lightyellow-orange",
zlim = c(-0.5, 0.5)
)
}
\arguments{
\item{data}{an input vector containing gene symbols}

\item{xTarget}{a "dTarget" or "sTarget" object with the componet
'list_pNode' related to 'eGene' predictors. Alternatively, it can be a
data frame with columns
('context','mode','probeID','Symbol','gene_cse','snps','snp_cse','A1','A2','b_GWAS','b_eQTL','b_ABF','p_GWAS','p_eQTL','pp_ABF','H4')}

\item{type}{the type of the heatmap. It can be "Gene" (gene-centric
heatmap) or "Gene_SNP" (heatmap for the gene-snp pair)}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted}
}
\value{
an object of class "ggplot" appended with 'mat' (the matrix colored by
'b_ABF') and 'df' (a data frame with columns
'priority','code','context','mode','ProbeID','Symbol','gene_cse','snps','snp_cse','A1','A2','b_GWAS','b_eQTL','b_ABF','p_GWAS','p_eQTL','pp_ABF','direction_GWAS','direction_eQTL','direction_ABF').
}
\description{
\code{xPierABFheatmap} is supposed to visualise ABF evidence using
heatmap. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xPierABFheatmap(data, dTarget)
}
}
\seealso{
\code{\link{xSparseMatrix}}, \code{\link{xHeatmap}},
\code{\link{xSparseMatrix}}
}
