% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierCor.r
\name{xPierCor}
\alias{xPierCor}
\title{Function to calculate correlation between prioritised genes and user-defined external data}
\usage{
xPierCor(pNode, data, method = c("pearson", "spearman"),
pvalue.type = c("nominal", "empirical"), seed = 825, nperm = 2000,
plot = FALSE)
}
\arguments{
\item{pNode}{an object of class "pNode" (or "sTarget" or "dTarget").
Alternatively, it can be a data frame with two columns ('name' and
'priority')}

\item{data}{a named input vector containing numeric data for genes
(gene symbols)}

\item{method}{the method used to calcualte correlation. It can be
'pearson' for Pearson's correlation or 'spearman' for Spearman rank
correlation}

\item{pvalue.type}{the type of the p-value calcualted. It can be
'nominal' for nominal p-value or 'empirical' for empirical p-value}

\item{seed}{an integer specifying the seed}

\item{nperm}{the number of random permutations}

\item{plot}{logical to indicate whether scatter plot is drawn}
}
\value{
an object of class "ggplot" with a component 'res' (a data frame with
two columns 'cor' and 'pval') if the scatter plot is drawn. Otherwise
it is a data frame with two columns 'cor' and 'pval'.
}
\description{
\code{xPierCor} is supposed to calculate correlation between
prioritised genes and user-defined external data.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
# a) provide the seed nodes/genes with the weight info
## load ImmunoBase
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
## get genes within 500kb away from AS GWAS lead SNPs
seeds.genes <- ImmunoBase$AS$genes_variants
## seeds weighted according to distance away from lead SNPs
data <- 1- seeds.genes/500000

# b) perform priority analysis
pNode <- xPierGenes(data=data, network="PCommonsDN_medium",restart=0.7,
RData.location=RData.location)

# c) do correlation
data <- pNode$priority$priority[1:100]
name(data) <- pNode$priority$name[1:100]
res <- xPierCor(pNode, data, method="pearson", pvalue.type="empirical",
nperm=2000, plot=TRUE)
}
}
\seealso{
\code{\link{xPierCor}}
}
