% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierCross.r
\name{xPierCross}
\alias{xPierCross}
\title{Function to extract priority matrix from a list of dTarget/sTarget objects}
\usage{
xPierCross(list_xTarget, displayBy = c("rating", "rank", "pvalue",
"fdr"), combineBy = c("intersect", "union"), aggregateBy = c("none",
"fishers", "logistic", "Ztransform", "orderStatistic"), verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{list_xTarget}{a list of "dTarget"/"sTarget" objects or a
"dTarget"/"sTarget" object}

\item{displayBy}{which priority will be extracted. It can be "rating"
for priority score/rating (by default), "rank" for priority rank,
"pvalue" for priority p-value, "fdr" for priority fdr}

\item{combineBy}{how to resolve nodes/targets from a list of
"dTarget"/"sTarget" objects. It can be "intersect" for intersecting
nodes (by default), "union" for unionising nodes}

\item{aggregateBy}{the aggregate method used. It can be either "none"
for no aggregation, or "orderStatistic" for the method based on the
order statistics of p-values, "fishers" for Fisher's method,
"Ztransform" for Z-transform method, "logistic" for the logistic
method. Without loss of generality, the Z-transform method does well in
problems where evidence against the combined null is spread widely
(equal footings) or when the total evidence is weak; Fisher's method
does best in problems where the evidence is concentrated in a
relatively small fraction of the individual tests or when the evidence
is at least moderately strong; the logistic method provides a
compromise between these two. Notably, the aggregate methods 'fishers'
and 'logistic' are preferred here}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
If aggregateBy is 'none' (by default), a data frame containing priority
matrix, with each column/disease for either priority score/rating, or
priorty rank or priority p-value.
If aggregateBy is not 'none', an object of the class "cTarget", a list
with following components:
\itemize{
\item{\code{priority}: a data frame of nGene X 6 containing gene
priority (aggregated) information, where nGene is the number of genes,
and the 6 columns are "name" (gene names), "rank" (ranks of the
priority scores), "pvalue" (the aggregated p-value, converted from
empirical cumulative distribution of the probability of being GSP),
"fdr" (fdr adjusted from the aggregated p-value), "priority"
(-log10(pvalue) but rescaled into the 5-star ratings), "description"
(gene description)}
\item{\code{disease}: a data frame containing disease matrix, with each
column/disease for either priority score, or priorty rank or priority
p-value}
}
}
\description{
\code{xPierCross} is supposed to extract priority matrix from a list of
dTarget objects. Also supported is the aggregation of priority matrix
(similar to the meta-analysis) generating the priority results; we view
this functionality as the cross mode of the prioritisation.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
df_score <- xPierCross(ls_xTarget)
}
}
\seealso{
\code{\link{xPierMatrix}}
}
