% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierEvidence.r
\name{xPierEvidence}
\alias{xPierEvidence}
\title{Function to extract evidence from a list of pNode objects}
\usage{
xPierEvidence(list_pNode, target.query = NULL, verbose = TRUE)
}
\arguments{
\item{list_pNode}{a list of "pNode" objects or a "pNode" object}

\item{target.query}{which gene is in query. If NULL, all genes will be
queried}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
a data frame of nPair X 5 containing Gene-SNP pair info per context,
where the 5 columns are "Gene" (seed genes), "SNP" (dbSNP), "Score" (an
SNP's genetic influential score on a seed gene), "Context"
(predictors), "Flag" (indicative of Lead SNPs or LD SNPs)
}
\description{
\code{xPierEvidence} is supposed to extract evidence from a list of
pNode objects, in terms of seed genes under genetic influence.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
df_Gene2SNP <- xPierEvidence(ls_pNode)
}
}
\seealso{
\code{\link{xPierSNPsAdv}}
}
