% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierKEGG.r
\name{xPierKEGG}
\alias{xPierKEGG}
\title{Function to visualise prioritised genes in terms of a KEGG pathway}
\usage{
xPierKEGG(xTarget, vis = c("net", "evidence", "pathview"), hsa =
"hsa04621",
priority.top = NULL, incoming.neighbor.order = 1, nodes_query = NULL,
largest.comp = TRUE, pathview.filename = NULL,
pathview.filetype = c("png", "pdf"), verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata", ...)
}
\arguments{
\item{xTarget}{an object of class "dTarget" or "sTarget"}

\item{vis}{the type of visualisation for a KEGG pathway. It can be one
of "net" (visualising the network with nodes colored by priority score;
by default), "evidence" for visualising the network with nodes as pie
charts, and "pathview" for using the package "pathview"}

\item{hsa}{the identity of KEGG pathway in query. The full list of
pathways in human can be found at
\url{http://www.genome.jp/kegg-bin/show_organism?menu_type=pathway_maps&org=hsa}.
For example, 'hsa04621' for 'NOD-like receptor signaling pathway',
where the prefix 'hsa' can be ignored}

\item{priority.top}{the number of the top targets. By default, it is
NULL meaning no such restriction}

\item{incoming.neighbor.order}{an integer giving the order of the
incoming neighborhood. By default, it is 1-order incoming neighborhood}

\item{nodes_query}{which gene in query will be visualised. It (if not
null) has the high priority over nodes selected by 'priority.top' and
'incoming.neighbor.order' above}

\item{largest.comp}{logical to indicate whether the largest component
is only retained. By default, it sets to true for the largest component
being left}

\item{pathview.filename}{the file name saved using the package
"pathview". By default, it is NULL meaning "hsa.Pi"}

\item{pathview.filetype}{the file format saved using the package
"pathview". It can be "png" or "pdf"}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}

\item{...}{additional graphic parameters. If the type of visualisation
is 'net', see \code{\link{xVisNet}}; if the visualisation type is
'evidence', see \code{\link{xVisEvidence}}}
}
\value{
a subgraph, an object of class "igraph".
}
\description{
\code{xPierKEGG} is supposed to visualise prioritised genes in terms of
a KEGG pathway. It returns an object of class "igraph".
}
\note{
If vis is 'pathview', it will depend on whether a package "pathview"
has been installed. It can be installed via:
\code{source("http://bioconductor.org/biocLite.R");
biocLite("pathview")}.
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
## evidence
xPierKEGG(xTarget, hsa="hsa04621", vis="evidence",
RData.location=RData.location)
## network
xPierKEGG(xTarget, hsa="hsa04621", vis="net",
RData.location=RData.location)
## using pathview
pv.out <- xPierKEGG(xTarget, hsa="hsa04621", vis="pathview",
pathview.filetype=c("png","pdf")[2], RData.location=RData.location)
}
}
\seealso{
\code{\link{xVisNet}}, \code{\link{xVisEvidence}}
}
