% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierManhattan.r
\name{xPierManhattan}
\alias{xPierManhattan}
\title{Function to visualise prioritised genes using manhattan plot}
\usage{
xPierManhattan(pNode, color = c("darkred", "darkgreen"), top = 50,
top.label.type = c("box", "text"), top.label.size = 2,
top.label.col = "darkblue", y.scale = c("normal", "sqrt"),
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"), verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{pNode}{an object of class "pNode"}

\item{color}{a character vector for colors to alternate chromosome
colorings. If NULL, ggplot2 default colors will be used. If a single
character is provided, it can be "jet" (jet colormap) or "rainbow"
(rainbow colormap, that is, red-yellow-green-cyan-blue-magenta)}

\item{top}{the number of the top targets to be labelled/highlighted}

\item{top.label.type}{how to label the top targets. It can be "box"
drawing a box around the labels , and "text" for the text only}

\item{top.label.size}{the highlight label size}

\item{top.label.col}{the highlight label color}

\item{y.scale}{how to transform the y scale. It can be "normal" for no
transformation, and "sqrt" for square transformation}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location"}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
an object of class "ggplot", appended by an GR object called 'gr'
}
\description{
\code{xPierManhattan} is supposed to visualise prioritised genes using
manhattan plot. Genes with the top priority are highlighed. It returns
an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

\dontrun{
# b) perform priority analysis
pNode <- xPierSNPs(data=AS, include.eQTL="JKng_mono",
include.HiC='Monocytes', network="PCommonsUN_medium", restart=0.7,
RData.location=RData.location)

# c) manhattan plot
## default plot
mp <- xPierManhattan(pNode, RData.location=RData.location)
#pdf(file="Gene_manhattan.pdf", height=6, width=12, compress=TRUE)
print(mp)
#dev.off()
mp$gr
## control visuals
mp <- xPierManhattan(pNode, color='ggplot2', top=50,
top.label.col="black", y.scale="sqrt", RData.location=RData.location)
mp
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xPier}},
\code{\link{xPierSNPs}}, \code{\link{xPierGenes}},
\code{\link{xPierPathways}}
}

