% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierMatrix.r
\name{xPierMatrix}
\alias{xPierMatrix}
\title{Function to extract priority matrix from a list of pNode objects}
\usage{
xPierMatrix(list_pNode, displayBy = c("score", "rank"),
combineBy = c("intersect", "union"), verbose = TRUE)
}
\arguments{
\item{list_pNode}{a list of "pNode" objects}

\item{displayBy}{which priority will be extracted. It can be "score"
for priority score (by default), "rank" for priority rank}

\item{combineBy}{how to resolve nodes/targets from a list of "pNode"
objects. It can be "intersect" for intersecting nodes (by default),
"union" for unionising nodes}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
a data frame containing priority matrix.
}
\description{
\code{xPierMatrix} is supposed to extract priority matrix from a list
of pNode objects
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
df_score <- xPierMatrix(ls_pNode)
}
}
\seealso{
\code{\link{xPierSNPs}}
}

