% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierROCR.r
\name{xPierROCR}
\alias{xPierROCR}
\title{Function to assess the dTarget performance via ROC and Precision-Recall (PR) analysis}
\usage{
xPierROCR(dTarget, GSP, GSN, verbose = TRUE)
}
\arguments{
\item{dTarget}{a data frame containing dTargets along with predictive
scores. It has two columns: 1st column for target, 2nd column for
predictive scores (the higher the better). Alternatively, it can be an
object of class "pNode" (or "sTarget" or "dTarget") from which a data
frame is extracted}

\item{GSP}{a vector containing Gold Standard Positives (GSP)}

\item{GSN}{a vector containing Gold Standard Negatives (GSN)}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to TRUE for display}
}
\value{
an object of the class "dTarget", a list with following components:
\itemize{
\item{\code{priority}: a data frame of nGene X 7 containing gene
priority (aggregated) information, where nGene is the number of genes,
and the 7 columns are "GS" (either 'GSP', or 'GSN', or 'NEW'), "name"
(gene names), "rank" (ranks of the priority scores), "pvalue" (the
aggregated p-value, converted from empirical cumulative distribution of
the probability of being GSP), "fdr" (fdr adjusted from the aggregated
p-value), "priority" (-log10(pvalue) but rescaled into the 5-star
ratings), "description" (gene description) and seed info including
"Overall" for the number of different types of seeds, followed by
details on individual type of seeds (that is, "OMIM", "Phenotype",
"Function", "nearbyGenes", "eQTL", "HiC")}
\item{\code{predictor}: a data frame containing predictor matrix, with
each column/predictor for either priority score, or priorty rank or
priority p-value}
\item{\code{metag}: an "igraph" object}
\item{\code{pPerf}: a "pPerf" object, with components "PRS", "AUROC",
"Fmax", "ROC_perf", "PR_perf", "Pred_obj"}
}
}
\description{
\code{xPierROCR} is supposed to assess the dTarget performance via
Receiver Operating Characteristic (ROC) and Precision-Recall (PR)
analysis. It requires three inputs: 1) Gold Standard Positive (GSP)
targets; 2) Gold Standard Negative (GSN) targets; 3) dTarget containing
predicted targets and predictive scores.
}
\note{
AUC: the area under ROC
F-measure: the maximum of a harmonic mean between precision and recall
along PR curve
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
dTarget <- xPierROCR(dTarget, GSP, GSN)
gp <- xPredictCompare(dTarget$pPerf)
}
}
