% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierSMRheatmap.r
\name{xPierSMRheatmap}
\alias{xPierSMRheatmap}
\title{Function to visualise SMR evidence using heatmap}
\usage{
xPierSMRheatmap(data, xTarget, colormap = "steelblue-white-darkred",
zlim = c(-1, 1))
}
\arguments{
\item{data}{an input vector containing gene symbols}

\item{xTarget}{a "dTarget" or "sTarget" object with the componet
'list_pNode' related to 'eGene' predictors. Alternatively, it can be a
data frame with columns
('Context','mode','probeID','Gene','ProbeChr','Probe_bp','topSNP','topSNP_chr','topSNP_bp','A1','A2','b_GWAS','b_eQTL','b_SMR','p_GWAS','p_eQTL','p_SMR','fdr_SMR')}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), and
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta). Alternatively, any
hyphen-separated HTML color names, e.g. "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{zlim}{the minimum and maximum z values for which colors should be
plotted}
}
\value{
an object of class "ggplot" appended with 'mat' (the matrix colored by
'b_SMR') and 'df' (a data frame with columns
'priority','code','Context','mode','probeID','Gene','ProbeChr','Probe_bp','topSNP','topSNP_chr','topSNP_bp','A1','A2','b_GWAS','b_eQTL','b_SMR','p_GWAS','p_eQTL','p_SMR','fdr_SMR','direction_GWAS','direction_eQTL','direction_SMR').
Only those fdr_SMR<0.05 are visualised.
}
\description{
\code{xPierSMRheatmap} is supposed to visualise SMR evidence using
heatmap. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
RData.location <- "http://galahad.well.ox.ac.uk/bigdata/"

gp <- xPierSMRheatmap(data, dTarget)
}
}
\seealso{
\code{\link{xPierSMRheatmap}}
}
