% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierTrackAdv.r
\name{xPierTrackAdv}
\alias{xPierTrackAdv}
\title{Function to visualise a list of prioritised genes using advanced track plot}
\usage{
xPierTrackAdv(pNode, priority.top = NULL, targets.query = NULL,
window = 1e+06, nearby = NULL, query.highlight = TRUE,
track.ideogram = TRUE, track.genomeaxis = TRUE,
name.datatrack = "Priority index", name.annotrack = "Genes",
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"), SNPs = NULL,
GR.SNP = c("dbSNP_GWAS", "dbSNP_Common", "dbSNP_Single"),
verbose = TRUE,
RData.location = "http://galahad.well.ox.ac.uk/bigdata", ...)
}
\arguments{
\item{pNode}{an object of class "pNode" (or "sTarget" or "dTarget")}

\item{priority.top}{the number of the top targets used for track plot.
By default, it is NULL meaning all targets are used}

\item{targets.query}{which genes in query will be visualised. If NULL,
the target gene with the top priority will be displayed}

\item{window}{the maximum distance defining nearby genes around the
target gene in query. By default it is 1e6}

\item{nearby}{the maximum number defining nearby genes around the
target gene in query. By default it is NULL. If not NULL, it will
overwrite the parameter 'window'}

\item{query.highlight}{logical to indicate whether the gene in query
will be highlighted}

\item{track.ideogram}{logical to indicate whether ideogram track is
shown. By default, it is TRUE}

\item{track.genomeaxis}{logical to indicate whether genome axis track
is shown. By default, it is TRUE}

\item{name.datatrack}{the name for the data track. By default, it is
"Priority index"}

\item{name.annotrack}{the name for the annotation track. By default, it
is "Target genes"}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location"}

\item{SNPs}{a input vector containing SNPs. SNPs should be provided as
dbSNP ID (ie starting with rs). Alternatively, they can be in the
format of 'chrN:xxx', where N is either 1-22 or X, xxx is genomic
positional number; for example, 'chr16:28525386'. By default, it is NLL
meaning the SNP annotation track will be not displayed}

\item{GR.SNP}{the genomic regions of SNPs. By default, it is
'dbSNP_GWAS', that is, SNPs from dbSNP (version 146) restricted to GWAS
SNPs and their LD SNPs (hg19). It can be 'dbSNP_Common', that is,
Common SNPs from dbSNP (version 146) plus GWAS SNPs and their LD SNPs
(hg19). Alternatively, the user can specify the customised input. To do
so, first save your RData file (containing an GR object) into your
local computer, and make sure the GR object content names refer to
dbSNP IDs. Then, tell "GR.SNP" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}

\item{...}{additional graphic parameters. For example, the parameter
"strip" allows the panel title is hided (FALSE), shown (TRUE) or
without the background (lattice::strip.custom(bg="transparent")); the
parameter "layout" allows specification of the layout (the first
element for the columns and the second element for the rows). See
\url{http://www.rdocumentation.org/packages/lattice/topics/xyplot} for
the complete list.}
}
\value{
an object of class "trellis"
}
\description{
\code{xPierTrackAdv} is supposed to visualise prioritised genes using
advanced track plot. Internally, it calls the function 'xPierTrack' per
gene.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

# b) perform priority analysis
pNode <- xPierSNPs(data=AS, include.eQTL="JKng_mono",
include.HiC='Monocytes', network="PCommonsUN_medium", restart=0.7,
RData.location=RData.location)

# c) track plot
library(Gviz)
#pdf(file="Gene_tracks.pdf", height=4, width=10, compress=TRUE)
xPierTrackAdv(pNode, RData.location=RData.location)
#dev.off()
xPierTrackAdv(pNode, priority.top=1000, nearby=20,
RData.location=RData.location)
}
}
\seealso{
\code{\link{xMLrandomforest}}
}
