% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPredictCompare.r
\name{xPredictCompare}
\alias{xPredictCompare}
\title{Function to compare prediction performance results}
\usage{
xPredictCompare(list_pPerf, displayBy = c("ROC", "PR"), sort = TRUE,
detail = TRUE, facet = FALSE, signature = TRUE)
}
\arguments{
\item{list_pPerf}{a list of "pPerf" objects}

\item{displayBy}{which curve will be used for comparison. It can be
"ROC" for ROC curve (by default), "PR" for PR curve}

\item{sort}{logical to indicate whether to sort methods according to
performance. By default, it sets TRUE}

\item{detail}{logical to indicate whether to label methods along with
performance. By default, it sets TRUE}

\item{facet}{logical to indicate whether to facet/wrap a 1d of panels
into 2d. By default, it sets FALSE}

\item{signature}{a logical to indicate whether the signature is
assigned to the plot caption. By default, it sets TRUE showing which
function is used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xPredictCompare} is supposed to compare prediction performance
results. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
# Load the library
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
bp <- xPredictCompare(ls_pPerf, displayBy="ROC")
print(bp)
## modify legend position
bp + theme(legend.position=c(0.75,0.25))
}
}
\seealso{
\code{\link{xPredictROCR}}
}

