% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xSNP2eGenes.r
\name{xSNP2eGenes}
\alias{xSNP2eGenes}
\title{Function to define eQTL genes given a list of SNPs or a customised eQTL mapping data}
\usage{
xSNP2eGenes(data, include.eQTL = c(NA, "JKscience_TS2A",
"JKscience_TS2B",
"JKscience_TS3A", "JKng_bcell", "JKng_mono", "JKnc_neutro", "JK_nk",
"GTEx_V4_Adipose_Subcutaneous", "GTEx_V4_Artery_Aorta",
"GTEx_V4_Artery_Tibial", "GTEx_V4_Esophagus_Mucosa",
"GTEx_V4_Esophagus_Muscularis", "GTEx_V4_Heart_Left_Ventricle",
"GTEx_V4_Lung", "GTEx_V4_Muscle_Skeletal", "GTEx_V4_Nerve_Tibial",
"GTEx_V4_Skin_Sun_Exposed_Lower_leg", "GTEx_V4_Stomach",
"GTEx_V4_Thyroid",
"GTEx_V4_Whole_Blood"), eQTL.customised = NULL,
cdf.function = c("empirical", "exponential"), plot = F, verbose = T,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/Portal")
}
\arguments{
\item{data}{a input vector containing SNPs. SNPs should be provided as
dbSNP ID (ie starting with rs). Alternatively, they can be in the
format of 'chrN:xxx', where N is either 1-22 or X, xxx is number; for
example, 'chr16:28525386'}

\item{include.eQTL}{genes modulated by eQTL (also Lead SNPs or in LD
with Lead SNPs) are also included. By default, it is 'NA' to disable
this option. Otherwise, those genes modulated by eQTL will be included:
immune stimulation in monocytes ('JKscience_TS1A' and 'JKscience_TS2B'
for cis-eQTLs or 'JKscience_TS3A' for trans-eQTLs) from Science 2014,
343(6175):1246949; cis- and trans-eQTLs in B cells ('JKng_bcell') and
in monocytes ('JKng_mono') from Nature Genetics 2012, 44(5):502-510;
cis- and trans-eQTLs in neutrophils ('JKnc_neutro') from Nature
Communications 2015, 7(6):7545; cis-eQTLs in NK cells ('JK_nk') which
is unpublished. Also supported are GTEx cis-eQTLs from Science 2015,
348(6235):648-60, including 13 tissues:
'GTEx_Adipose_Subcutaneous','GTEx_Artery_Aorta','GTEx_Artery_Tibial','GTEx_Esophagus_Mucosa','GTEx_Esophagus_Muscularis','GTEx_Heart_Left_Ventricle','GTEx_Lung','GTEx_Muscle_Skeletal','GTEx_Nerve_Tibial','GTEx_Skin_Sun_Exposed_Lower_leg','GTEx_Stomach','GTEx_Thyroid','GTEx_Whole_Blood'.}

\item{eQTL.customised}{a user-input matrix or data frame with 3
columns: 1st column for SNPs/eQTLs, 2nd column for Genes, and 3rd for
eQTL mapping significance level (p-values or FDR). It is designed to
allow the user analysing their eQTL data. This customisation (if
provided) has the high priority over built-in eQTL data.}

\item{cdf.function}{a character specifying a Cumulative Distribution
Function (cdf). It can be one of 'exponential' based on exponential
cdf, 'empirical' for empirical cdf}

\item{plot}{logical to indicate whether the histogram plot (plus
density or CDF plot) should be drawn. By default, it sets to false for
no plotting}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{Gene}: eQTL genes}
\item{\code{SNP}: eQTLs}
\item{\code{Pval}: the eQTL mapping significant level}
\item{\code{Weight}: the eQTL weight}
}
}
\description{
\code{xSNP2eGenes} is supposed to define eQTL genes given a list of
SNPs or a customised eQTL mapping data. The eQTL weight is calcualted
as Cumulative Distribution Function of negative log-transformed
eQTL-reported signficance level.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(Pi)

# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)

# b) define eQTL genes
df_eGenes <- xSNP2eGenes(data=AS$SNP,
include.eQTL=c("JKscience_TS2A","JKscience_TS2B","JKscience_TS3A"))
}
}
\seealso{
\code{\link{xRDataLoader}}
}

