% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSNPeqtl.r
\name{xSNPeqtl}
\alias{xSNPeqtl}
\title{Function to extract eQTL-gene pairs given a list of SNPs or a customised eQTL mapping data}
\usage{
xSNPeqtl(data = NULL, include.eQTL = c(NA, "JKscience_TS2A",
"JKscience_TS2B", "JKscience_TS3A", "JKng_bcell", "JKng_mono",
"JKnc_neutro",
"JK_nk", "GTEx_V4_Adipose_Subcutaneous", "GTEx_V4_Artery_Aorta",
"GTEx_V4_Artery_Tibial", "GTEx_V4_Esophagus_Mucosa",
"GTEx_V4_Esophagus_Muscularis", "GTEx_V4_Heart_Left_Ventricle",
"GTEx_V4_Lung", "GTEx_V4_Muscle_Skeletal", "GTEx_V4_Nerve_Tibial",
"GTEx_V4_Skin_Sun_Exposed_Lower_leg", "GTEx_V4_Stomach",
"GTEx_V4_Thyroid",
"GTEx_V4_Whole_Blood", "eQTLdb_NK", "eQTLdb_CD14", "eQTLdb_LPS2",
"eQTLdb_LPS24", "eQTLdb_IFN"), eQTL.customised = NULL,
verbose = TRUE, RData.location =
"http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{data}{NULL or a input vector containing SNPs. If NULL, all SNPs
will be considered. If a input vector containing SNPs, SNPs should be
provided as dbSNP ID (ie starting with rs). Alternatively, they can be
in the format of 'chrN:xxx', where N is either 1-22 or X, xxx is
number; for example, 'chr16:28525386'}

\item{include.eQTL}{genes modulated by eQTL (also Lead SNPs or in LD
with Lead SNPs) are also included. By default, it is 'NA' to disable
this option. Otherwise, those genes modulated by eQTL will be included:
immune stimulation in monocytes ('JKscience_TS1A' and 'JKscience_TS2B'
for cis-eQTLs or 'JKscience_TS3A' for trans-eQTLs) from Science 2014,
343(6175):1246949; cis- and trans-eQTLs in B cells ('JKng_bcell') and
in monocytes ('JKng_mono') from Nature Genetics 2012, 44(5):502-510;
cis- and trans-eQTLs in neutrophils ('JKnc_neutro') from Nature
Communications 2015, 7(6):7545; cis-eQTLs in NK cells ('JK_nk') which
is unpublished. Also supported are GTEx cis-eQTLs from Science 2015,
348(6235):648-60, including 13 tissues:
"GTEx_V4_Adipose_Subcutaneous","GTEx_V4_Artery_Aorta","GTEx_V4_Artery_Tibial","GTEx_V4_Esophagus_Mucosa","GTEx_V4_Esophagus_Muscularis","GTEx_V4_Heart_Left_Ventricle","GTEx_V4_Lung","GTEx_V4_Muscle_Skeletal","GTEx_V4_Nerve_Tibial","GTEx_V4_Skin_Sun_Exposed_Lower_leg","GTEx_V4_Stomach","GTEx_V4_Thyroid","GTEx_V4_Whole_Blood".}

\item{eQTL.customised}{a user-input matrix or data frame with 3
columns: 1st column for SNPs/eQTLs, 2nd column for Genes, and 3rd for
eQTL mapping significance level (p-values or FDR). It is designed to
allow the user analysing their eQTL data. This customisation (if
provided) has the high priority over built-in eQTL data.}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{SNP}: eQTLs}
\item{\code{Gene}: eQTL-containing genes}
\item{\code{Sig}: the eQTL mapping significant level}
\item{\code{Context}: the context in which eQTL data was generated}
}
}
\description{
\code{xSNPeqtl} is supposed to extract eQTL-gene pairs given a list of
SNPs or a customised eQTL mapping data.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase')
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

# b) define eQTL genes
df_SGS <- xSNPeqtl(data=AS[,1], include.eQTL="JKscience_TS2A")
}
\seealso{
\code{\link{xRDataLoader}}
}

