% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSNPhic.r
\name{xSNPhic}
\alias{xSNPhic}
\title{Function to extract promoter capture HiC-gene pairs given a list of SNPs}
\usage{
xSNPhic(data = NULL, entity = c("SNP", "chr:start-end", "data.frame",
"bed",
"GRanges"), include.HiC = c(NA, "Monocytes", "Macrophages_M0",
"Macrophages_M1", "Macrophages_M2", "Neutrophils", "Megakaryocytes",
"Endothelial_precursors", "Erythroblasts", "Fetal_thymus",
"Naive_CD4_T_cells", "Total_CD4_T_cells",
"Activated_total_CD4_T_cells",
"Nonactivated_total_CD4_T_cells", "Naive_CD8_T_cells",
"Total_CD8_T_cells",
"Naive_B_cells", "Total_B_cells"), GR.SNP = c("dbSNP_GWAS",
"dbSNP_Common"),
verbose = TRUE, RData.location =
"http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{data}{NULL or a input vector containing SNPs. If NULL, all SNPs
will be considered. If a input vector containing SNPs, SNPs should be
provided as dbSNP ID (ie starting with rs) or in the format of
'chrN:xxx', where N is either 1-22 or X, xxx is number; for example,
'chr16:28525386'. Alternatively, it can be other formats/entities (see
the next parameter 'entity')}

\item{entity}{the data entity. By default, it is "SNP". For general
use, it can also be one of "chr:start-end", "data.frame", "bed" or
"GRanges"}

\item{include.HiC}{genes linked to input SNPs are also included. By
default, it is 'NA' to disable this option. Genes linked to input SNPs
are based on Promoter Capture HiC (PCHic), including 17 primary blood
cell types:
"Monocytes","Macrophages_M0","Macrophages_M1","Macrophages_M2","Neutrophils","Megakaryocytes","Endothelial_precursors","Erythroblasts","Fetal_thymus","Naive_CD4_T_cells","Total_CD4_T_cells","Activated_total_CD4_T_cells","Nonactivated_total_CD4_T_cells","Naive_CD8_T_cells","Total_CD8_T_cells","Naive_B_cells","Total_B_cells".}

\item{GR.SNP}{the genomic regions of SNPs. By default, it is
'dbSNP_GWAS', that is, SNPs from dbSNP (version 146) restricted to GWAS
SNPs and their LD SNPs (hg19). It can be 'dbSNP_Common', that is,
Common SNPs from dbSNP (version 146) plus GWAS SNPs and their LD SNPs
(hg19). Alternatively, the user can specify the customised input. To do
so, first save your RData file (containing an GR object) into your
local computer, and make sure the GR object content names refer to
dbSNP IDs. Then, tell "GR.SNP" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location". Note: you can also load your customised GR object
directly}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
If input data is NULL, a data frame with following columns:
\itemize{
\item{\code{from}: baited genomic regions (baits)}
\item{\code{to}: preyed (other end) genomic regions of interactions
(preys)}
\item{\code{score}: CHiCAGO scores quantifying the strength of physical
interactions between harbors and partners}
}
If input data is not NULL, a list with two components: "df" and "ig".
"df" is a data frame with following columns:
\itemize{
\item{\code{from}: 'from/bait' genomic regions}
\item{\code{to}: 'to/prey' genomic regions}
\item{\code{score}: CHiCAGO scores quantifying the strength of physical
interactions between baits and preys}
\item{\code{from_genes}: genes associated with 'from/bait' genomic
regions}
\item{\code{to_genes}: genes associated with 'to/prey' genomic regions}
\item{\code{SNP}: input SNPs (in query)}
\item{\code{SNP_end}: specify which end SNPs in query fall into (either
'bait/from' or 'prey/to')}
\item{\code{SNP_harbor}: genomic regions harbors the SNPs in query}
\item{\code{Context}: the context in which PCHiC data was generated}
}
"ig" is an object of both classes "igraph" and "PCHiC", a direct graph
with nodes for genomic regions and edges for CHiCAGO scores between
them. Also added node attribute is 1) 'target' storing genes assocated
and 2) 'SNP' for input SNPs (if the node harboring input SNPs). If
several cell types are queried, "ig" is actually a list of
"igraph"/"PCHiC" objects.
}
\description{
\code{xSNPhic} is supposed to extract HiC-gene pairs given a list of
SNPs.
}
\note{
None
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
data <- names(ImmunoBase$AS$variants)

\dontrun{
# b) extract HiC-gene pairs given a list of AS SNPs
PCHiC <- xSNPhic(data, include.HiC="Monocytes", GR.SNP="dbSNP_GWAS",
RData.location=RData.location)
head(PCHiC$df)

# c) visualise the interaction (a directed graph: bait->prey)
g <- PCHiC$ig
## a node with SNPs colored in 'skyblue' and the one without SNPs in 'pink'
## the width in an edge is proportional to the interaction strength
xPCHiCplot(g, vertex.label.cex=0.5)
xPCHiCplot(g, glayout=layout_in_circle, vertex.label.cex=0.5)
}
}
\seealso{
\code{\link{xRDataLoader}}
}

