% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xVisEvidence.r
\name{xVisEvidence}
\alias{xVisEvidence}
\title{Function to visualise evidence for prioritised genes in a gene network}
\usage{
xVisEvidence(dTarget, g = NA, nodes = NULL, neighbor.order = 1,
neighbor.seed = TRUE, colormap = "ggplot2", legend.position =
"topleft",
verbose = TRUE, ...)
}
\arguments{
\item{dTarget}{an object of class "dTarget"}

\item{g}{an object of class "igraph". If NA, the 'metag' will be used,
which is part of the object of class "dTarget"}

\item{nodes}{which node genes are in query. If NULL, the top gene will
be queried}

\item{neighbor.order}{an integer giving the order of the neighborhood.
By default, it is 1-order neighborhood}

\item{neighbor.seed}{logical to indicate whether neighbors are seeds
only. By default, it sets to true}

\item{colormap}{short name for the colormap. It can be one of "jet"
(jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red
colormap), "wyr" (white-yellow-red colormap), "br" (black-red
colormap), "yr" (yellow-red colormap), "wb" (white-black colormap),
"rainbow" (rainbow colormap, that is,
red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2
default color palette). Alternatively, any hyphen-separated HTML color
names, e.g. "lightyellow-orange" (by default), "blue-black-yellow",
"royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of
standard color names can be found in
\url{http://html-color-codes.info/color-names}}

\item{legend.position}{the legend position. If NA, the legend will be
hiden}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{...}{additional graphic parameters. See
\url{http://igraph.org/r/doc/plot.common.html} for the complete list.}
}
\value{
a subgraph, an object of class "igraph".
}
\description{
\code{xVisEvidence} is supposed to visualise evidence for prioritised
genes in a gene network. It returns an object of class "igraph".
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
subg <- xVisEvidence(dTarget, nodes="B2M", neighbor.order=1,
neighbor.seed=TRUE, vertex.label.color="black", vertex.label.cex=1,
vertex.label.dist=1, legend.position="bottomleft", signature=FALSE)
}
}
\seealso{
\code{\link{xPierMatrix}}
}
