% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGSEAdotplot.r
\name{xGSEAdotplot}
\alias{xGSEAdotplot}
\title{Function to visualise GSEA results using dot plot}
\usage{
xGSEAdotplot(eGSEA, top = 1, priority.color = c("lightyellow",
"orange"),
peak = TRUE, compact = FALSE, signature = TRUE)
}
\arguments{
\item{eGSEA}{an object of class "eGSEA"}

\item{top}{the number of the top enrichments to be visualised.
Alternatively, the gene set names can be queried}

\item{priority.color}{a character vector for coloring priority scores}

\item{peak}{logical to indicate whether the peak info of leading genes
is labelled}

\item{compact}{logical to indicate whether the compact/void theme is
used. If TRUE, axes and legend info will be hidden}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot" or a list of "ggplot" objects.
}
\description{
\code{xGSEAdotplot} is supposed to visualise GSEA results using dot
plot. It returns an object of class "ggplot" or a list of "ggplot"
objects.
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
\dontrun{
gp <- xGSEAdotplot(eGSEA, top=1)
#gp <- xGSEAdotplot(eGSEA, top=1, peak=FALSE, compact=TRUE, signature=FALSE)
gp

ls_gp <- xGSEAdotplot(eGSEA, top=1:4, signature=FALSE)
library(gridExtra)
grid.arrange(grobs=ls_gp, ncol=2)
}
}
\seealso{
\code{\link{xPierGSEA}}
}
