% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xMLdensity.r
\name{xMLdensity}
\alias{xMLdensity}
\title{Function to visualise machine learning results using density plot}
\usage{
xMLdensity(xTarget, displayBy = c("All", "GS", "GSN", "GSP", "NEW"),
x.scale = c("sqrt", "normal"), font.family = "sans",
signature = TRUE)
}
\arguments{
\item{xTarget}{an object of class "xTarget" or "dTarget" (with the
component 'pPerf')}

\item{displayBy}{which targets will be used for displaying. It can be
one of "GS" for gold standard targets, "GSN" for gold standard
negatives, "GSP" for gold standard positives, "NEW" for putative/new
targets (non-GS), "All" for all targets (by default)}

\item{x.scale}{how to transform the x scale. It can be "normal" for no
transformation, and "sqrt" for square root transformation (by default)}

\item{font.family}{the font family for texts}

\item{signature}{logical to indicate whether the signature is assigned
to the plot caption. By default, it sets TRUE showing which function is
used to draw this graph}
}
\value{
an object of class "ggplot"
}
\description{
\code{xMLdensity} is supposed to visualise machine learning results
using density plot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
gp <- xMLdensity(xTarget, displayBy="All")
gp
}
}
\seealso{
\code{\link{xMLrandomforest}}
}
