% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPierManhattan.r
\name{xPierManhattan}
\alias{xPierManhattan}
\title{Function to visualise prioritised genes using manhattan plot}
\usage{
xPierManhattan(pNode, color = c("darkred", "darkgreen"), cex = 0.5,
highlight.top = 20, highlight.col = "deepskyblue",
highlight.label.size = 2, highlight.label.offset = 0.02,
highlight.label.col = "darkblue", y.scale = c("normal", "sqrt",
"log10"),
GR.Gene = c("UCSC_knownGene", "UCSC_knownCanonical"), verbose = TRUE,
RData.location =
"https://github.com/hfang-bristol/RDataCentre/blob/master/Portal")
}
\arguments{
\item{pNode}{an object of class "pNode"}

\item{color}{a character vector for point colors to alternate}

\item{cex}{a numeric value for point size}

\item{highlight.top}{the number of the top targets to be highlighted}

\item{highlight.col}{the highlight colors}

\item{highlight.label.size}{the highlight label size}

\item{highlight.label.offset}{the highlight label offset}

\item{highlight.label.col}{the highlight label color}

\item{y.scale}{how to transform the y scale. It can be "normal" for no
transformation, "sqrt" for square transformation, and "log10" for
10-based log-transformation}

\item{GR.Gene}{the genomic regions of genes. By default, it is
'UCSC_knownGene', that is, UCSC known genes (together with genomic
locations) based on human genome assembly hg19. It can be
'UCSC_knownCanonical', that is, UCSC known canonical genes (together
with genomic locations) based on human genome assembly hg19.
Alternatively, the user can specify the customised input. To do so,
first save your RData file (containing an GR object) into your local
computer, and make sure the GR object content names refer to Gene
Symbols. Then, tell "GR.Gene" with your RData file name (with or
without extension), plus specify your file RData path in
"RData.location"}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to false for no display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
an object of class "ggplot"
}
\description{
\code{xPierManhattan} is supposed to visualise prioritised genes using
manhattan plot. Genes with the top priority are highlighed. It returns
an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase')
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

\dontrun{
# b) perform priority analysis
pNode <- xPierSNPs(data=AS, network="PCommonsUN_medium",restart=0.7)

# c) manhattan plot
mp <- xPierManhattan(pNode, highlight.top=10)
#pdf(file="Gene_manhattan.pdf", height=6, width=12, compress=TRUE)
print(mp)
#dev.off()
}
}
\seealso{
\code{\link{xRDataLoader}}, \code{\link{xPier}},
\code{\link{xPierSNPs}}, \code{\link{xPierGenes}},
\code{\link{xPierPathways}}
}

