% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xPredictCompare.r
\name{xPredictCompare}
\alias{xPredictCompare}
\title{Function to compare prediction performance results}
\usage{
xPredictCompare(
list_pPerf,
displayBy = c("ROC", "PR"),
type = c("curve", "bar"),
sort = TRUE,
detail = TRUE,
facet = FALSE,
font.family = "sans",
signature = TRUE
)
}
\arguments{
\item{list_pPerf}{a list of "pPerf" objects}

\item{displayBy}{which performance will be used for comparison. It can
be "ROC" for ROC curve (by default), "PR" for PR curve}

\item{type}{the type of plot to draw. It can be "curve" for curve plot
(by default), "bar" for bar plot}

\item{sort}{logical to indicate whether to sort methods according to
performance. By default, it sets TRUE}

\item{detail}{logical to indicate whether to label methods along with
performance. By default, it sets TRUE}

\item{facet}{logical to indicate whether to facet/wrap a 1d of panels
into 2d. By default, it sets FALSE}

\item{font.family}{the font family for texts}

\item{signature}{a logical to indicate whether the signature is
assigned to the plot caption. By default, it sets TRUE showing which
function is used to draw this graph}
}
\value{
an object of class "ggplot" or NULL (if all input pPerf objects are
NULL)
}
\description{
\code{xPredictCompare} is supposed to compare prediction performance
results. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
RData.location <- "http://galahad.well.ox.ac.uk/bigdata"
\dontrun{
bp <- xPredictCompare(ls_pPerf, displayBy="ROC")
print(bp)
## modify legend position
bp + theme(legend.position=c(0.75,0.25))
}
}
\seealso{
\code{\link{xPredictCompare}}
}
