% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSNP2eGenes.r
\name{xSNP2eGenes}
\alias{xSNP2eGenes}
\title{Function to define eQTL genes given a list of SNPs or a customised eQTL mapping data}
\usage{
xSNP2eGenes(data, include.eQTL = c(NA, "JKscience_TS2A",
"JKscience_TS2A_CD14", "JKscience_TS2A_LPS2", "JKscience_TS2A_LPS24",
"JKscience_TS2A_IFN", "JKscience_TS2B", "JKscience_TS2B_CD14",
"JKscience_TS2B_LPS2", "JKscience_TS2B_LPS24", "JKscience_TS2B_IFN",
"JKscience_TS3A", "JKng_bcell", "JKng_bcell_cis", "JKng_bcell_trans",
"JKng_mono", "JKng_mono_cis", "JKng_mono_trans", "JKnc_neutro",
"JKnc_neutro_cis", "JKnc_neutro_trans", "JK_nk",
"GTEx_V4_Adipose_Subcutaneous", "GTEx_V4_Artery_Aorta",
"GTEx_V4_Artery_Tibial", "GTEx_V4_Esophagus_Mucosa",
"GTEx_V4_Esophagus_Muscularis", "GTEx_V4_Heart_Left_Ventricle",
"GTEx_V4_Lung", "GTEx_V4_Muscle_Skeletal", "GTEx_V4_Nerve_Tibial",
"GTEx_V4_Skin_Sun_Exposed_Lower_leg", "GTEx_V4_Stomach",
"GTEx_V4_Thyroid",
"GTEx_V4_Whole_Blood", "eQTLdb_NK", "eQTLdb_CD14", "eQTLdb_LPS2",
"eQTLdb_LPS24", "eQTLdb_IFN"), eQTL.customised = NULL,
cdf.function = c("empirical", "exponential"), plot = FALSE,
verbose = TRUE, RData.location =
"http://galahad.well.ox.ac.uk/bigdata")
}
\arguments{
\item{data}{a input vector containing SNPs. SNPs should be provided as
dbSNP ID (ie starting with rs). Alternatively, they can be in the
format of 'chrN:xxx', where N is either 1-22 or X, xxx is number; for
example, 'chr16:28525386'}

\item{include.eQTL}{genes modulated by eQTL (also Lead SNPs or in LD
with Lead SNPs) are also included. By default, it is 'NA' to disable
this option. Otherwise, those genes modulated by eQTL will be included.
Pre-built eQTL datasets are detailed in the section 'Note'}

\item{eQTL.customised}{a user-input matrix or data frame with 3
columns: 1st column for SNPs/eQTLs, 2nd column for Genes, and 3rd for
eQTL mapping significance level (p-values or FDR). It is designed to
allow the user analysing their eQTL data. This customisation (if
provided) has the high priority over built-in eQTL data.}

\item{cdf.function}{a character specifying a Cumulative Distribution
Function (cdf). It can be one of 'exponential' based on exponential
cdf, 'empirical' for empirical cdf}

\item{plot}{logical to indicate whether the histogram plot (plus
density or CDF plot) should be drawn. By default, it sets to false for
no plotting}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}

\item{RData.location}{the characters to tell the location of built-in
RData files. See \code{\link{xRDataLoader}} for details}
}
\value{
a data frame with following columns:
\itemize{
\item{\code{Gene}: eQTL-containing genes}
\item{\code{SNP}: eQTLs}
\item{\code{Sig}: the eQTL mapping significant level (the
best/minimum)}
\item{\code{Weight}: the eQTL weight}
}
}
\description{
\code{xSNP2eGenes} is supposed to define eQTL genes given a list of
SNPs or a customised eQTL mapping data. The eQTL weight is calcualted
as Cumulative Distribution Function of negative log-transformed
eQTL-reported signficance level.
}
\note{
Pre-built eQTL datasets are described below according to the data
sources.\cr
1. Context-specific eQTLs in monocytes: resting and activating states.
Sourced from Science 2014, 343(6175):1246949
\itemize{
\item{\code{JKscience_TS2A}: cis-eQTLs in either state (based on 228
individuals with expression data available for all experimental
conditions).}
\item{\code{JKscience_TS2A_CD14}: cis-eQTLs only in the resting/CD14+
state (based on 228 individuals).}
\item{\code{JKscience_TS2A_LPS2}: cis-eQTLs only in the activating
state induced by 2-hour LPS (based on 228 individuals).}
\item{\code{JKscience_TS2A_LPS24}: cis-eQTLs only in the activating
state induced by 24-hour LPS (based on 228 individuals).}
\item{\code{JKscience_TS2A_IFN}: cis-eQTLs only in the activating state
induced by 24-hour interferon-gamma (based on 228 individuals).}
\item{\code{JKscience_TS2B}: cis-eQTLs in either state (based on 432
individuals).}
\item{\code{JKscience_TS2B_CD14}: cis-eQTLs only in the resting/CD14+
state (based on 432 individuals).}
\item{\code{JKscience_TS2B_LPS2}: cis-eQTLs only in the activating
state induced by 2-hour LPS (based on 432 individuals).}
\item{\code{JKscience_TS2B_LPS24}: cis-eQTLs only in the activating
state induced by 24-hour LPS (based on 432 individuals).}
\item{\code{JKscience_TS2B_IFN}: cis-eQTLs only in the activating state
induced by 24-hour interferon-gamma (based on 432 individuals).}
\item{\code{JKscience_TS3A}: trans-eQTLs in either state.}
}
2. eQTLs in B cells. Sourced from Nature Genetics 2012, 44(5):502-510
\itemize{
\item{\code{JKng_bcell}: cis- and trans-eQTLs.}
\item{\code{JKng_bcell_cis}: cis-eQTLs only.}
\item{\code{JKng_bcell_trans}: trans-eQTLs only.}
}
3. eQTLs in monocytes. Sourced from Nature Genetics 2012, 44(5):502-510
\itemize{
\item{\code{JKng_mono}: cis- and trans-eQTLs.}
\item{\code{JKng_mono_cis}: cis-eQTLs only.}
\item{\code{JKng_mono_trans}: trans-eQTLs only.}
}
4. eQTLs in neutrophils. Sourced from Nature Communications 2015,
7(6):7545
\itemize{
\item{\code{JKnc_neutro}: cis- and trans-eQTLs.}
\item{\code{JKnc_neutro_cis}: cis-eQTLs only.}
\item{\code{JKnc_neutro_trans}: trans-eQTLs only.}
}
5. eQTLs in NK cells. Unpublished
\itemize{
\item{\code{JK_nk}: cis-eQTLs.}
}
6. Tissue-specific eQTLs from GTEx (version 4; incuding 13 tissues).
Sourced from Science 2015, 348(6235):648-60
\itemize{
\item{\code{GTEx_V4_Adipose_Subcutaneous}: cis-eQTLs in tissue 'Adipose
Subcutaneous'.}
\item{\code{GTEx_V4_Artery_Aorta}: cis-eQTLs in tissue 'Artery Aorta'.}
\item{\code{GTEx_V4_Artery_Tibial}: cis-eQTLs in tissue 'Artery
Tibial'.}
\item{\code{GTEx_V4_Esophagus_Mucosa}: cis-eQTLs in tissue 'Esophagus
Mucosa'.}
\item{\code{GTEx_V4_Esophagus_Muscularis}: cis-eQTLs in tissue
'Esophagus Muscularis'.}
\item{\code{GTEx_V4_Heart_Left_Ventricle}: cis-eQTLs in tissue 'Heart
Left Ventricle'.}
\item{\code{GTEx_V4_Lung}: cis-eQTLs in tissue 'Lung'.}
\item{\code{GTEx_V4_Muscle_Skeletal}: cis-eQTLs in tissue 'Muscle
Skeletal'.}
\item{\code{GTEx_V4_Nerve_Tibial}: cis-eQTLs in tissue 'Nerve Tibial'.}
\item{\code{GTEx_V4_Skin_Sun_Exposed_Lower_leg}: cis-eQTLs in tissue
'Skin Sun Exposed Lower leg'.}
\item{\code{GTEx_V4_Stomach}: cis-eQTLs in tissue 'Stomach'.}
\item{\code{GTEx_V4_Thyroid}: cis-eQTLs in tissue 'Thyroid'.}
\item{\code{GTEx_V4_Whole_Blood}: cis-eQTLs in tissue 'Whole Blood'.}
}
}
\examples{
\dontrun{
# Load the library
library(Pi)
}

RData.location <- "http://galahad.well.ox.ac.uk/bigdata_dev"
# a) provide the SNPs with the significance info
## get lead SNPs reported in AS GWAS and their significance info (p-values)
#data.file <- "http://galahad.well.ox.ac.uk/bigdata/AS.txt"
#AS <- read.delim(data.file, header=TRUE, stringsAsFactors=FALSE)
ImmunoBase <- xRDataLoader(RData.customised='ImmunoBase',
RData.location=RData.location)
gr <- ImmunoBase$AS$variants
AS <- as.data.frame(GenomicRanges::mcols(gr)[, c('Variant','Pvalue')])

\dontrun{
# b) define eQTL genes
df_eGenes <- xSNP2eGenes(data=AS[,1], include.eQTL="JKscience_TS2A",
RData.location=RData.location)
}
}
\seealso{
\code{\link{xRDataLoader}}
}

